+++ packages/cli/rollup.config.js
@@ -7,7 +7,8 @@ export default config({
     './src/commands/create-app/create-app.ts',
     './src/commands/gql-serve/gql-serve.ts',
     './src/commands/cms-sync-s3/cms-sync-s3.ts',
-    './src/commands/init-s3-sync/init-s3-sync.ts'
+    './src/commands/init-s3-sync/init-s3-sync.ts',
+    './src/commands/gen-fragments/gen-fragments.ts'
   ],
   babelHelpers: 'runtime'
 });