+++ packages/component-library/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @last-rev/component-library
 
+## 0.1.23
+
+### Patch Changes
+
+- 5336e17: Fix Hero height and Navigation mobile interaction
+
+## 0.1.22
+
+### Patch Changes
+
+- 0fdd5a6: Fix navigation item
+
 ## 0.1.21
 
 ### Patch Changes